+++ xc/programs/xlockmore/modes/matrix.c
@@ -8,7 +8,7 @@ static const char sccsid[] = "@(#)matrix.c	4.15 99/06/30 xlockmore";
 /* Matrix-style screensaver
  *
  * Author: Erik O'Shaughnessy (eriko@xenolab.com)  20 Apr 1999
- * 
+ *
  * Permission to use, copy, modify, and distribute this software and its
  * documentation for any purpose and without fee is hereby granted,
  * provided that the above copyright notice appear in all copies and that
@@ -29,7 +29,7 @@ static const char sccsid[] = "@(#)matrix.c	4.15 99/06/30 xlockmore";
  *              - fix memory-related bugs found by xlock maintainer
  *              - properly reinitialize screen with new size when init()
  *                is called multiple times
- *              - reinstate multiple final characters (suggested by 
+ *              - reinstate multiple final characters (suggested by
  *                Joan Touzet)
  *              - reduce delay to 1 ms (provisionally)
  *              - decouple pixmap height from screen height, to permit
@@ -38,7 +38,7 @@ static const char sccsid[] = "@(#)matrix.c	4.15 99/06/30 xlockmore";
  *                will help if the machine is busy.  Also restores
  *                broken display on very small windows).
  *              - nevertheless, certain display parameters such as the
- *                interval between column speed updates should depend on 
+ *                interval between column speed updates should depend on
  *                the screen height -- enforce this at runtime.
  *              - fix column spacing to use entire window
  *              - remove redundant size fields from matrix_t and column_t
@@ -51,7 +51,7 @@ static const char sccsid[] = "@(#)matrix.c	4.15 99/06/30 xlockmore";
  * 27-Jun-1999: Overhauled by Jeremy Buhler (jbuhler@cs.washington.edu)
  *              - use screen-to-screen copies as much as possible
  *              - get by with a single pixmap per column instead of two
- *              - try to optimize draw_matrix() and new_column() as much 
+ *              - try to optimize draw_matrix() and new_column() as much
  *                as humanly possible
  *              - fixed double-freeing in release_matrix()
  *              - increased delay to 10 ms to prevent hyperfast scrolling
@@ -91,8 +91,9 @@ ModeSpecOpt matrix_opts = {0, NULL, 0, NULL, NULL};
 #ifdef USE_MODULES
 ModStruct   matrix_description =
 {"matrix", "init_matrix", "draw_matrix", "release_matrix",
- "refresh_matrix", "change_matrix", NULL, &matrix_opts, 
- 1000, 1, 1, 1, 64, 1.0, "", "Shows the Matrix", 0, NULL};
+ "refresh_matrix", "change_matrix", NULL, &matrix_opts,
+ 1000, 1, 1, 1, 64, 1.0, "",
+ "Shows the Matrix", 0, NULL};
 #endif
 
 
@@ -102,14 +103,14 @@ ModStruct   matrix_description =
 typedef struct {
 	int yPtr;		/* current offset relative to pixmap origin */
 	Pixmap pixmap;
-  
+
 	int strLen;		/* remaining length of current string */
-	int gapLen;		/* remaining length of current gap    */ 
+	int gapLen;		/* remaining length of current gap    */
 	Bool endChar;		/* is the next char an ending char?   */
 
 	int speed;		/* pixels downward per update */
 	int nextSpeedUpdate;	/* time until next change in speed */
-  
+
 } column_t;
 
 
@@ -120,30 +121,30 @@ typedef struct {
 	column_t *columns;      /* columns of the Matrix display */
 	int num_cols;		/* number of columns of Matrix data
 				 * (depends on screen width) */
- 
+
 	int charsPerPixmap;	/* height of column pixmaps in characters
 				 * (depends on screen height) */
-  
+
 	int pixmapHeight;	/* height of column pixmaps in pixels
                                  * (depends on screen height) */
-	 
+
 	int speedUpdateInterval; /* number of characters or spaces that
 				 * must be written to column pixmap before
 				 * the column speed is changed
 				 * (depends on screen height) */
-  
-	Pixmap kana[2];		/* pixmap containing katakana 
+
+	Pixmap kana[2];		/* pixmap containing katakana
 				 * [0] is fg on bg
 				 * [1] is bold on bg */
-  
+
 	Pixel fg;		/* foreground */
 	Pixel bg;		/* background */
 	Pixel bold;		/* standout foreground color */
 } matrix_t;
 
 
-/* These Katakana font bitmaps were taken from the file 12x24rk.bdf 
- * distributed with XFree86 3.3.3.  The numbers were taken from the file 
+/* These Katakana font bitmaps were taken from the file 12x24rk.bdf
+ * distributed with XFree86 3.3.3.  The numbers were taken from the file
  * 10x20.bdf, since the 12x24rk numbers were rather large compared to
  * the other characters.
  */
@@ -330,7 +331,7 @@ static const unsigned char katakana_bits[] = {
  * slower hardware).  However, it can be done in relatively little memory
  * because the pixmaps can be kept short and regenerated frequently.
  *
- * Undefining RANDOMIZE_COLUMNS will cause the pixmaps to be generated only 
+ * Undefining RANDOMIZE_COLUMNS will cause the pixmaps to be generated only
  * once, which should save a lot of horsepower for whatever else your machine
  * might be doing.  However, the tradeoff is that the pixmaps are longer
  * to avoid a really boring display.
@@ -342,7 +343,7 @@ static const unsigned char katakana_bits[] = {
  * Must be <= katakana_cell_height; interesting values range
  * from one (crawl) to about fifteen (zoom!).
  * (Note: the distribution of speeds implemented in new_column()
- *  is slightly biased against very slow scrolling.) 
+ *  is slightly biased against very slow scrolling.)
  */
 #define MATRIX_SPEED  MATRIX_RANDOM(1, 8)
 
@@ -379,10 +380,10 @@ static void new_column(const ModeInfo *mi, column_t *);
 
 /* NAME: init_matrix
  *
- * FUNCTION: allocate space for global matrix array 
+ * FUNCTION: allocate space for global matrix array
  *           initialize colors
  *           initialize dimensions
- *           allocate a pair of pixmaps containing character set 
+ *           allocate a pair of pixmaps containing character set
  *               (0 is green on black, 1 is bold on black)
  *           create columns of "matrix" data
  */
@@ -390,7 +391,7 @@ static void new_column(const ModeInfo *mi, column_t *);
 void init_matrix(ModeInfo *mi)
 {
   matrix_t *mp;
-  
+
   if (matrix == NULL)
 	if((matrix = (matrix_t *)
 		calloc(MI_NUM_SCREENS(mi), sizeof(matrix_t))) == NULL)
@@ -398,12 +399,12 @@ void init_matrix(ModeInfo *mi)
 		perror("init_matrix:calloc");
 		return;
 	  }
-  
+
   /* don't want any exposure events from XCopyArea */
   XSetGraphicsExposures(MI_DISPLAY(mi), MI_GC(mi), False);
-  
+
   mp = PICK_MATRIX(mi);
-  
+
   /* screen may have changed size/depth/who-knows-what */
   free_screen(mi, mp);
   init_active_screen(mi);
@@ -419,9 +420,9 @@ void draw_matrix(ModeInfo *mi)
   matrix_t *mp = PICK_MATRIX(mi);
   double xDistance = (double) MI_WIDTH(mi) / (double) mp->num_cols;
   int i;
-  
+
   MI_IS_DRAWN(mi) = True;
-  
+
   /*
    * THEORY OF OPERATION
    *
@@ -437,39 +438,39 @@ void draw_matrix(ModeInfo *mi)
    * call to draw_matrix(); when it reaches zero, the column's pixmap
    * may be redrawn to generate new characters for the column.
    *
-   * Second, the scroll increment corresponding to <speed> need not be a 
-   * multiple of the pixmap height. We must special-case the handling of the 
+   * Second, the scroll increment corresponding to <speed> need not be a
+   * multiple of the pixmap height. We must special-case the handling of the
    * last decrement that makes yPtr <= 0 to ensure that the last few pixels
    * of the old data AND the first few pixels of the new data are both drawn.
-   * It turns out to be quite reasonable to do the pixmap update inline, 
+   * It turns out to be quite reasonable to do the pixmap update inline,
    * between drawing the last few old pixels and the first few new pixels.
    * Hence, we can achieve smooth scrolling with only one pixmap per column.
    */
-  
-  for (i = 0; i < mp->num_cols; i++) 
+
+  for (i = 0; i < mp->num_cols; i++)
 	{
 	  column_t *c = &(mp->columns[i]);
 	  int xOffset = (int) (i * xDistance);
 	  int yDelta  = c->speed;
-	  
+
 	  /* screen-screen blit takes care of most of the column */
-	  XCopyArea(MI_DISPLAY(mi), MI_WINDOW(mi), 
+	  XCopyArea(MI_DISPLAY(mi), MI_WINDOW(mi),
 				MI_WINDOW(mi), MI_GC(mi),
 				xOffset, 0,
 				PIXMAP_WIDTH, MI_HEIGHT(mi) - yDelta,
 				xOffset, yDelta);
-	  
+
 	  c->yPtr -= yDelta;
-	  
+
 	  if (c->yPtr <= 0) /* we're about to finish the current pixmap */
 		{
 		  /* exhaust the pixels in the current pixmap */
 		  XCopyArea(MI_DISPLAY(mi), c->pixmap,
 					MI_WINDOW(mi), MI_GC(mi),
-					0, 0, 
+					0, 0,
 					PIXMAP_WIDTH, yDelta + c->yPtr,
 					xOffset, -(c->yPtr));
-		  
+
 #ifdef RANDOMIZE_COLUMNS
 		  /* regenerate the column pixmap with new text */
 		  new_column(mi, c);
@@ -478,15 +479,15 @@ void draw_matrix(ModeInfo *mi)
 		   * so take this opportunity to update the column speed.
 		   */
 		  c->speed = MATRIX_SPEED;
-#endif	
-		  
+#endif
+
 		  /* fill in the remainder of the column from the new pixmap */
 		  XCopyArea(MI_DISPLAY(mi), c->pixmap,
 					MI_WINDOW(mi), MI_GC(mi),
-					0, mp->pixmapHeight + c->yPtr, 
+					0, mp->pixmapHeight + c->yPtr,
 					PIXMAP_WIDTH, -(c->yPtr),
 					xOffset, 0);
-		  
+
 		  c->yPtr += mp->pixmapHeight;
 		}
 	  else
@@ -494,7 +495,7 @@ void draw_matrix(ModeInfo *mi)
 		  /* fill in the remainder of the column from the pixmap */
 		  XCopyArea(MI_DISPLAY(mi), c->pixmap,
 					MI_WINDOW(mi), MI_GC(mi),
-					0, c->yPtr, 
+					0, c->yPtr,
 					PIXMAP_WIDTH, yDelta,
 					xOffset, 0);
 		}
@@ -510,15 +511,15 @@ void draw_matrix(ModeInfo *mi)
 void release_matrix(ModeInfo *mi)
 {
   int screen;
-  
+
   /* If the matrix exists, free all data associated with all screens.
    * free_screen() does no harm if given nonexistent screens.
    */
   if (matrix != NULL)
 	{
-	  for (screen = 0; screen < MI_NUM_SCREENS(mi); screen++) 
+	  for (screen = 0; screen < MI_NUM_SCREENS(mi); screen++)
 		free_screen(mi, &matrix[screen]);
-	  
+
 	  (void) free((void *) matrix);
 	  matrix = NULL;
 	}
@@ -534,13 +535,13 @@ void refresh_matrix(ModeInfo *mi)
 {
   matrix_t *mp = PICK_MATRIX(mi);
   int c;
-  
+
   /* We simply clear the whole window and restart the scrolling operation.
    * In principle, it is possible to restore at least some of the
    * column pixmap data to the screen, but it's not worth the trouble.
    */
   MI_CLEARWINDOW(mi);
-  
+
   for (c = 0; c < mp->num_cols; c++)
 	mp->columns[c].yPtr = mp->pixmapHeight;
 }
@@ -555,18 +556,18 @@ void change_matrix(ModeInfo *mi)
 {
   matrix_t *mp = PICK_MATRIX(mi);
   int c;
-  
+
   MI_CLEARWINDOW(mi);
-  
+
   for (c = 0; c < mp->num_cols; c++)
 	{
 	  column_t *column = &(mp->columns[c]);
-	  
+
 	  column->yPtr = mp->pixmapHeight;
-	  
+
 	  column->nextSpeedUpdate = 0; /* will generate new speed  */
 	  column->strLen  = 0;         /* will generate new string */
-	  
+
 	  new_column(mi, column);
 	}
 }
@@ -585,37 +586,37 @@ static void init_active_screen(const ModeInfo *mi)
 {
   matrix_t *mp = PICK_MATRIX(mi);
   int c;
- 
-  if (MI_NPIXELS(mi) > 2) 
+
+  if (MI_NPIXELS(mi) > 2)
 	{
 	  mp->fg   = MI_PIXEL(mi, GREEN);
 	  mp->bold = MI_PIXEL(mi, BRIGHTGREEN);
-	} 
-  else 
+	}
+  else
 	mp->fg = mp->bold = MI_WHITE_PIXEL(mi);
-  
+
   mp->bg = MI_BLACK_PIXEL(mi);
-  
-  mp->kana[0] = 
+
+  mp->kana[0] =
 	XCreatePixmapFromBitmapData(MI_DISPLAY(mi),
 								MI_WINDOW(mi), (char *) katakana_bits,
 								katakana_width, katakana_height,
 								mp->bg, mp->fg, MI_DEPTH(mi));
-  
-  mp->kana[1] = 
+
+  mp->kana[1] =
 	XCreatePixmapFromBitmapData(MI_DISPLAY(mi),
 								MI_WINDOW(mi), (char *) katakana_bits,
 								katakana_width, katakana_height,
 								mp->bg, mp->bold, MI_DEPTH(mi));
-  
-  
-  /* 
+
+
+  /*
    * Using the width of the kana font, determine how many
    * columns can fit across the current screen (with a little
    * padding).
    */
-  mp->num_cols = MI_WIDTH(mi) / (1.3 * katakana_cell_width);
-  
+  mp->num_cols = (int) (MI_WIDTH(mi) / (1.3 * katakana_cell_width));
+
   mp->columns = (column_t *) calloc(mp->num_cols, sizeof(column_t));
   if (mp->columns == NULL)
 	{
@@ -629,34 +630,34 @@ static void init_active_screen(const ModeInfo *mi)
    * make the height large to maximize variety without updating the pixmaps.
    */
 #ifdef RANDOMIZE_COLUMNS
-  mp->charsPerPixmap = MI_HEIGHT(mi) / (4 * katakana_cell_height); 
+  mp->charsPerPixmap = MI_HEIGHT(mi) / (4 * katakana_cell_height);
 #else
-  mp->charsPerPixmap = MI_HEIGHT(mi) / katakana_cell_height; 
+  mp->charsPerPixmap = MI_HEIGHT(mi) / katakana_cell_height;
 #endif
-  
+
   if (!mp->charsPerPixmap)  /* sanity check for tiny windows */
 	mp->charsPerPixmap = 1;
-  
+
   mp->pixmapHeight = mp->charsPerPixmap * katakana_cell_height;
-  
+
   for (c = 0; c < mp->num_cols; c++)
 	mp->columns[c].pixmap  = XCreatePixmap(MI_DISPLAY(mi), MI_WINDOW(mi),
 										   PIXMAP_WIDTH, mp->pixmapHeight,
 										   MI_DEPTH(mi));
-  
+
   /* Change the speed of a column after it's covered half the screen */
   mp->speedUpdateInterval = MI_HEIGHT(mi) / (2 * katakana_cell_height);
-  
+
   /* initialize the column data */
   for (c = 0; c < mp->num_cols; c++)
 	{
 	  column_t *column = &(mp->columns[c]);
-	  
+
 	  column->yPtr = mp->pixmapHeight;
-	  
+
 	  column->nextSpeedUpdate = 0; /* will generate new speed  */
 	  column->strLen  = 0;         /* will generate new string */
-	  
+
 	  new_column(mi, column);
 	}
 }
@@ -673,23 +674,23 @@ static void free_screen(const ModeInfo *mi, matrix_t *mp)
   if (mp->columns != NULL)
 	{
 	  int c;
-	  
+
 	  for (c = 0; c < mp->num_cols; c++)
 		{
 		  if (mp->columns[c].pixmap)
 			XFreePixmap(MI_DISPLAY(mi), mp->columns[c].pixmap);
 		}
-	  
+
 	  free(mp->columns);
 	  mp->columns = NULL;
 	}
-  
+
   if (mp->kana[0])
 	{
 	  XFreePixmap(MI_DISPLAY(mi), mp->kana[0]);
 	  mp->kana[0] = 0;
 	}
-  
+
   if (mp->kana[1])
 	{
 	  XFreePixmap(MI_DISPLAY(mi), mp->kana[1]);
@@ -707,15 +708,15 @@ static void new_column(const ModeInfo *mi, column_t *c)
 {
   matrix_t *mp = PICK_MATRIX(mi);
   int currChar;
-  
-  /* 
+
+  /*
    * clear the pixmap to get rid of previous data or initialize
    */
   XSetForeground(MI_DISPLAY(mi), MI_GC(mi), mp->bg);
   XFillRectangle(MI_DISPLAY(mi), c->pixmap, MI_GC(mi),
 				 0, 0, PIXMAP_WIDTH, mp->pixmapHeight);
-  
-  /* 
+
+  /*
    * Write characters into the column pixmap, starting at the
    * bottom and moving upwards.  Decrement the string length, gap length,
    * and next-speed-update timers for the column and reinitialize them
@@ -727,20 +728,20 @@ static void new_column(const ModeInfo *mi, column_t *c)
 		{
 		  c->speed = MATRIX_SPEED;
 		  c->nextSpeedUpdate = mp->speedUpdateInterval;
-		  
+
 		  /* tweak to prevent really slow columns from remaining slow */
-		  if (c->speed <= 2) 
+		  if (c->speed <= 2)
 			c->nextSpeedUpdate /= 2;
 		}
-	  
+
 	  if (c->gapLen > 0)
 		{
 		  c->gapLen--;
 		  continue;
 		}
-	  else if (c->strLen == 0) 
+	  else if (c->strLen == 0)
 		{
-		  /* 
+		  /*
 		   * generate a gap of random length in the string, followed
 		   * by a random number of characters.  Set the endChar flag
 		   * to indicate the end of a new string.
@@ -753,9 +754,9 @@ static void new_column(const ModeInfo *mi, column_t *c)
 		{
 		  Pixmap src;
 		  int kOffset;
-		  
-		  if (c->endChar) 
-			{  
+
+		  if (c->endChar)
+			{
 			  src = mp->kana[1]; /* a bold, non-numeric character */
 			  kOffset = MATRIX_RANDOM(10, katakana_num_cells);
 			  c->endChar = False;
@@ -763,12 +764,12 @@ static void new_column(const ModeInfo *mi, column_t *c)
 			  src = mp->kana[0];
 			  kOffset = MATRIX_RANDOM(0, katakana_num_cells);
 			}
-		  
+
 		  XCopyArea(MI_DISPLAY(mi), src, c->pixmap, MI_GC(mi),
 					kOffset * katakana_cell_width, 0,
 					katakana_cell_width, katakana_cell_height,
 					0, currChar * katakana_cell_height);
-		  
+
 		  c->strLen--;
 		}
 	}
